make SPyPanic a subclass of SPyError, so it's automatically caugth by main
